src/test/kotlin/year2024/Day06Test.kt
package year2024

import Utils.Companion.cases
import Utils.Companion.readTestFile
import org.assertj.core.api.Assertions.*
import org.junit.jupiter.params.*
import org.junit.jupiter.params.provider.*

class Day06Test {

    companion object {
        @JvmStatic
        fun data() = cases(
            arrayOf(
                Arguments.of("year2024/Day06_sample", 41L, 6L),
            ),
            Arguments.of("year2024/Day06_personal", 5239L, 1753L)
        )
    }

    @ParameterizedTest(name = "File {0}, part 1 answer = {1}, part 2 answer = {2}")
    @MethodSource("data")
    fun test(file: String, part1Answer: Long, part2Answer: Long) {
        val contents = readTestFile(file)
        assertThat(Day06().part1(contents)).isEqualTo(part1Answer)
        assertThat(Day06().part2(contents)).isEqualTo(part2Answer)
    }
}
